+++ CHANGELOG.md
@@ -28,6 +28,8 @@ For details about compatibility between different releases, see the **Commitment
 
 - When a gateway uplink message contains duplicate data uplinks, only the one with the highest RSSI are forwarded.
 - The HTTP port now allows HTTP/2 connections over cleartext (h2c).
+- `ttn-lw-stack ns-db migrate` command records the schema version and only performs migrations if on a newer version.
+  - Use the `--force` flag to force perform migrations.
 
 ### Deprecated